Created base classes and c00.py
